+++ benchmarks/TimeseriesExtractor.py
@@ -1,5 +1,7 @@
 import glob, os, shutil, tempfile
 
+import joblib
+
 from neurocaps.extraction import TimeseriesExtractor
 
 ROOT_DIR = os.path.dirname(__file__)
@@ -10,7 +12,21 @@ class Time_TimeseriesExtractor:
         self.tmp_dir = tempfile.TemporaryDirectory()
         self.bids_dir = os.path.join(self.tmp_dir.name, "tests", "data", "dset")
         self.pipeline_name = os.path.join("derivatives", "fmriprep_1.0.0", "fmriprep")
+        self.parcel_approach = os.path.join(self.tmp_dir.name, "HCPex_parcel_approach.pkl")
 
+        # Copy parcel pickle
+        shutil.copyfile(os.path.join(ROOT_DIR, "tests", "data", "HCPex_parcel_approach.pkl"), self.parcel_approach)
+
+        # Copy nii file
+        nii_file = os.path.join(self.tmp_dir.name, "HCPex.nii.gz")
+        shutil.copyfile(os.path.join(ROOT_DIR, "tests", "data", "HCPex.nii.gz"), nii_file)
+
+        # Change location of "maps"
+        parcel_approach = joblib.load(self.parcel_approach)
+        parcel_approach["Custom"]["maps"] = nii_file
+        joblib.dump(parcel_approach, self.parcel_approach)
+
+        # Copy data
         shutil.copytree(os.path.join(ROOT_DIR, "tests", "data", "dset"), self.bids_dir)
 
         work_dir = os.path.join(self.bids_dir, self.pipeline_name)
@@ -24,7 +40,7 @@ class Time_TimeseriesExtractor:
                 os.rename(file, file.replace("sub-01_", f"sub-{sub_id}_"))
 
     def time_get_bold(self):
-        extractor = TimeseriesExtractor()
+        extractor = TimeseriesExtractor(parcel_approach=self.parcel_approach)
 
         extractor.get_bold(bids_dir=self.bids_dir, pipeline_name=self.pipeline_name, task="rest", tr=1.2)